+++ backend/remote-state/azure/client_test.go
@@ -322,7 +322,7 @@ func TestRemoteClientWithKeyVaultEncryption(t *testing.T) {
 
 	ctx := context.TODO()
 	res := testResourceNamesWithKeyVault(rs, "testState", keyVaultName, keyName)
-	armClient := buildTestClientWithKeyVault(t, res)
+	armClient := buildTestClient(t, res)
 
 	err := armClient.buildTestResources(ctx, &res)
 	defer armClient.destroyTestResources(ctx, res)