+++ plans/changes_sync.go
@@ -21,6 +21,24 @@ type ChangesSync struct {
 	changes *Changes
 }
 
+// IsFullDestroy returns true if the set of changes indicates we are doing a
+// destroy of all resources.
+func (cs *ChangesSync) IsFullDestroy() bool {
+	if cs == nil {
+		panic("FullDestroy on nil ChangesSync")
+	}
+	cs.lock.Lock()
+	defer cs.lock.Unlock()
+
+	for _, c := range cs.changes.Resources {
+		if c.Action != Delete {
+			return false
+		}
+	}
+
+	return true
+}
+
 // AppendResourceInstanceChange records the given resource instance change in
 // the set of planned resource changes.
 //