+++ repl/format.go
@@ -16,6 +16,9 @@ func FormatValue(v cty.Value, indent int) string {
 	if !v.IsKnown() {
 		return "(known after apply)"
 	}
+	if v.IsMarked() {
+		return "(sensitive)"
+	}
 	if v.IsNull() {
 		ty := v.Type()
 		switch {